pyproject.toml
[tool.poetry]
authors = ["Marcel Hild <hild@b4mad.net>"]
description = ""
license = "GPL-3.0-or-later"
name = "mds-moderator"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
boto3 = "^1.35.23"
fastapi = "^0.115.0"
pipecat-ai = {extras = ["daily", "openai", "silero"], version = "==0.0.41"}
python = "^3.12"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
requests = "^2.32.3"
torch = { version = "==2.3.1", source = "pytorch" }
torchaudio =  { version = "==2.3.1", source = "pytorch" }
uvicorn = "^0.30.6"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pylance = "^0.18.0"
pytest = "^8.3.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
